feat: integration stat-service - main service, compilations - service, dao, controller

+++ ewm-service/src/main/java/ru/practicum/ewn/service/compilations/dto/CompilationUpdateDto.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Builder
 @Jacksonized
 public class CompilationUpdateDto {
-    List<Integer> events;
+    List<Long> events;
     Boolean pinned;
     String title;
 }
